refactor: move panel svg's with panels
